[+] reorganized code
