+++ src/data/vocabulary.py
@@ -4,13 +4,12 @@ import torch
 class Vocabulary:
     def __init__(self, labels: list[str] | str):
         """
-        Initializes a Vocabulary object with the given list of labels.
-        
         Args:
-            `labels` (`list`): A list of labels to be used in the vocabulary.
+            labels (list[str] | str): List of labels or a single label string.
         """
         self.idx2token = {0: "<PAD>", 1: '<START>',  2: "<END>"}
         self.token2idx = {"<PAD>": 0, "<START>": 1, "<END>": 2}
+        self.service_tokens = self.idx2token.copy()
         self.labels = labels
         self._build()
         del self.labels
@@ -24,26 +23,31 @@ class Vocabulary:
         self.idx2token.update({idx: char for idx, char in enumerate(self.labels, start_idx)})
         self.token2idx.update({char: idx for idx, char in enumerate(self.labels, start_idx)})
     
-    def encode(self, text: str):
+    def encode(self, text: str) -> torch.Tensor:
+        """Encodes a string into a sequence of tokens"""
         text = text.lower()
         for char in text:
             if char not in self.token2idx:
                 text = text.replace(char, '')
         # <START> ... <END>
-        output = [self.start_token_idx] + [self.token2idx[digit]for digit in text] + [self.end_token_idx]
+        output = [self.start_token_idx] + [self.token2idx[token] for token in text] + [self.end_token_idx]
         return torch.tensor(output)
     
-    def decode(self, tensor: torch.Tensor):
-        return ''.join([self.idx2token[idx] for idx in tensor.tolist()])
+    def decode(self, tensor: torch.Tensor) -> list[str]:
+        """Decodes a sequence of tokens into a list of characters"""
+        return [self.idx2token[idx] for idx in tensor.tolist()]
     
     @property
-    def pad_token_idx(self):
+    def pad_token_idx(self) -> int:
+        """Returns the index of the padding token"""
         return self.token2idx['<PAD>']
     
     @property
-    def start_token_idx(self):
+    def start_token_idx(self) -> int:
+        """Returns the index of the start token"""
         return self.token2idx['<START>']
     
     @property
-    def end_token_idx(self):
+    def end_token_idx(self) -> int:
+        """Returns the index of the end token"""
         return self.token2idx['<END>']
