+++ src/utils/__init__.py
@@ -1,0 +1,4 @@
+from .predictor import Predictor
+from .history import History
+
+__all__ = ['Predictor', 'History']
